Stream AggTrades end-to-end to avoid 2GB array limits
- PageManager.loadAggTrades: use ensureCached() + SQL COUNT instead of loading 13M trades into List
- PageManager.writeAggTradesData: stream from SQLite in 10k chunks with msgpack array header
- AggTradesStore: extract ensureCached() to fetch without returning full List
- AggTradesDao.countInRange: return long instead of int for large counts
- DataServiceClient.getAggTrades: use byteStream() instead of bytes() to avoid 2GB byte[]

+++ tradery-data-service/src/main/java/com/tradery/dataservice/data/sqlite/dao/AggTradesDao.java
@@ -371,7 +371,7 @@ public class AggTradesDao {
     /**
      * Count trades in a time range.
      */
-    public int countInRange(long startTime, long endTime) throws SQLException {
+    public long countInRange(long startTime, long endTime) throws SQLException {
         Connection c = conn.getConnection();
 
         String sql = "SELECT COUNT(*) FROM agg_trades WHERE timestamp >= ? AND timestamp <= ?";
@@ -379,10 +379,9 @@ public class AggTradesDao {
         try (PreparedStatement stmt = c.prepareStatement(sql)) {
             stmt.setLong(1, startTime);
             stmt.setLong(2, endTime);
-
             try (ResultSet rs = stmt.executeQuery()) {
                 if (rs.next()) {
-                    return rs.getInt(1);
+                    return rs.getLong(1);
                 }
             }
         }
